--- manifest.json
@@ -1,16 +1,0 @@
-{
-    "manifest_version": 3,
-    "name": "Comment Translator",
-    "version": "1.0",
-    "description": "Translates Youtube comments.",
-    "action": {
-        "default_title": "Comment Translator",
-        "default_popup": "html/popup.html"
-    },
-    "content_scripts": [
-        {
-            "js": ["js/content.js"],
-            "matches": ["https://www.youtube.com/watch?v=*"]
-        }
-    ]
-}
