+++ src/background/content.ts
