+++ src/components/Popup/popup.html
@@ -1,3 +1,4 @@
+<!-- File Restructure: Eventually delete this file. -->
 <!DOCTYPE html>
 <html lang="en">
 <head>
